[HBASE-10177] Fix the netty dependency issue
Author: gauravm

Summary:
The netty developers changed their group id from org.jboss.netty to io.netty. As a result, the zookeeper and hadoop dependencies pull in the older netty (3.2.2) and swift related dependencies pull in the newer netty (3.7.0). As a result we get ClassNotFoundExceptions, when the older 3.2.2 jar is picked up in place of 3.7.0.

This is the same as the commit made on the swift branch.

Test Plan: mvn dependency:tree doesn't pull in the 3.2.2 jar.

Reviewers: rshroff, liyintang, aaiyer, adela, fan

Reviewed By: adela

CC: hbase-dev@, daviddeng

Differential Revision: https://phabricator.fb.com/D1100077

Task ID: 3076680

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1551660 13f79535-47bb-0310-9956-ffa450edef68