[0.89-fb] [master] Ensure that rootRS death is processed correctly upon master failover.
Author: aaiyer

Summary:
We have seen cases where kill-hbase can get the system into a state where
the master is not able to assign any of the regions, because the master is
not assigning the root region.

This essentially seems to happen, because the old rootRegionServer was
being processed as dead. But, as part of that, the master does a root
scan -- that would fail because the region server is dead.

ProcessServerShutdown checks if the deadServer is the root server and tries
to get root assigned, when ProcessServerShutdown is instantiated. However, in
the case of a master failover, where the state is reconstructed from
ZKClusterStateRecovery, the rootRegionServer reference could be updated, after
the ProcessServerShutdown is created. In this case, PSS never makes any
progress after the split log stage.

Test Plan: re pro the scenario on my dev cluster.

Reviewers: liyintang, rshroff

Reviewed By: liyintang

CC: mbm, hbase-eng@

Differential Revision: https://phabricator.fb.com/D1091877

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1553891 13f79535-47bb-0310-9956-ffa450edef68